user roles set up

+++ app/controllers/wikis_controller.rb
@@ -6,14 +6,17 @@ class WikisController < ApplicationController
 
   def show
     @wiki =Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
     @wiki = current_user.wikis.new(wiki_params)
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki post saved."
@@ -27,10 +30,12 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.update_attributes(wiki_params)
       redirect_to @wiki
@@ -43,9 +48,10 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.destroy
-      flash[:notice] = "\"#{title}\" successfully deleted"
+      flash[:notice] = "Successfully deleted"
       redirect_to root_path
     else
       flash[:error] = "There was an error deleting this wiki post"
